fadl update

+++ YallaNotlob/app/controllers/orders_controller.rb
@@ -6,20 +6,31 @@ class OrdersController < ApplicationController
     # puts @orders.inspect
   end
 
+
   def finish
     @orders = Order.find params[:id]
+    if @orders.user_id  == current_user.id
       if @orders.status   = 'Finished'  
          @orders.save
          puts @orders.inspect
          redirect_to '/orders'
       end
+    else
+      redirect_to '/orders'
+    end
   end
+
+
 
   def cancel
     @orders = Order.find params[:id]
+    if @orders.user_id  == current_user.id
         if @orders.destroy
           redirect_to '/orders'
         end
+    else
+          redirect_to '/orders'   
+    end    
   end
 
   def new
